New (#22)
* autocomplete with errors

* 1

* added autocomplete

* ac

* separated cases into three components

* with debounce

* added debounce in script

* added debounce

* added debounce

* minor edits
