+++ App.vue
@@ -3,7 +3,6 @@
 </template>
 
 <script>
-import HelloWorld from './src/components/HelloWorld.vue'
 import { useStore } from 'vuex'
 import { onMounted, computed } from 'vue'
 import layoutPublic from './src/layouts/Public.vue'
@@ -12,7 +11,6 @@ import layoutSecure from './src/layouts/Secure.vue'
 export default {
   name: 'App',
   components: {
-    HelloWorld,
     layoutPublic,
     layoutSecure,
   },
@@ -21,7 +19,6 @@ export default {
     const storeUser = computed(() => store.state.user)
     onMounted(() => {
       console.log('signIn mounted')
-      // store.commit('login', 'defg')
     })
     return {
       storeUser,
