+++ main.js
@@ -3,6 +3,7 @@ import App from './App.vue'
 import './public/index.css'
 import router from './router.js'
 import store from './store.js'
+import './src/components/bwc-autocomplete.js'
 
 const app = createApp(App)
 app.use(router)
