+++ src/pages/LogIn.vue
@@ -1,27 +1,29 @@
 <template>
-  <div class="container">
-    <!-- <div class="sidebar"></div> -->
-    <div class="image">
-      <img src="../images/Group_116.png" />
-    </div>
-    <div class="form">
-      <div class="inputField">
-        <div class="inputTitle">Email ID</div>
-        <input class="input is-success" type="text" placeholder="Email ID" />
+  <div>
+    <Autocomplete />
+    <div class="container">
+      <div class="image">
+        <img src="../images/Group_116.png" />
       </div>
+      <div class="form">
+        <div class="input-Field">
+          <div class="input-title">Email ID</div>
+          <input class="input is-success" type="text" placeholder="Email ID" />
+        </div>
 
-      <div class="inputField">
-        <div class="inputTitle">Password</div>
-        <input class="input is-success" type="text" placeholder="Password" />
-      </div>
+        <div class="input-field">
+          <div class="input-title">Password</div>
+          <input class="input is-success" type="text" placeholder="Password" />
+        </div>
 
-      <div class="inputButton">
-        <button
-          @click="login"
-          class="button is-block is-link is-focused is-fullwidth"
-        >
-          Login
-        </button>
+        <div class="input-button">
+          <button
+            @click="login"
+            class="button is-block is-link is-focused is-fullwidth"
+          >
+            Log In
+          </button>
+        </div>
       </div>
     </div>
   </div>
@@ -49,20 +51,20 @@ export default {
 .form {
   text-align: left;
   font-size: 12;
-  padding-top: 300px;
+  padding-top: 270px;
   padding-bottom: 150px;
 }
 
-.inputField {
+.input-field {
   padding-top: 10px;
   padding-bottom: 10px;
 }
-.inputTitle {
+.input-title {
   font-weight: 400;
   color: black;
 }
 
-.inputButton {
+.input-button {
   padding-top: 10px;
   padding-bottom: 10px;
 }
@@ -74,20 +76,10 @@ export default {
 }
 
 .image {
-  padding-top: 50px;
+  padding-top: 20px;
+  padding-right: 20px;
   width: 240px;
   height: 180px;
   margin: right;
 }
-
-/* .sidebar {
-  height: 100%;
-  width: 100px;
-  position: fixed;
-  z-index: 1;
-  top: 0;
-  left: 0;
-  background-color: #08134b;
-  padding-top: 16px;
-} */
 </style>
